% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{add_spaces_to_lines}
\alias{add_spaces_to_lines}
\title{Add Spaces to the Beginning of Lines}
\usage{
add_spaces_to_lines(lines)
}
\arguments{
\item{lines}{A character vector containing the lines of text to be modified.}
}
\value{
A character vector with three spaces added to the beginning of each
line in the input vector.
}
\description{
This function adds three spaces to the beginning of each line in a character
vector. It processes each line by prefixing it with three spaces and returns
a modified character vector with the updated lines.
}
\examples{
lines <- c("Line 1", "Line 2", "Line 3")
result <- add_spaces_to_lines(lines)
# result will be c("   Line 1", "   Line 2", "   Line 3")
print(result)

}
