% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{add_string_to_first_line}
\alias{add_string_to_first_line}
\title{Add String to the First Line}
\usage{
add_string_to_first_line(lines, string_to_add)
}
\arguments{
\item{lines}{A character vector where the string will be added to the first
line.}

\item{string_to_add}{A character string that will be prepended to the first
line of the input vector.}
}
\value{
A character vector where the first line has the specified string
added to the beginning.
}
\description{
This function adds a specified string to the beginning of the first line of
a character vector. If the input vector is not empty, the function prepends
the provided string to the first element and returns the modified vector.
If the vector is empty, it returns the vector unchanged.
}
\examples{
lines <- c("First line", "Second line", "Third line")
result <- add_string_to_first_line(lines, "New Start: ")
# result will be c("New Start: First line", "Second line", "Third line")
print(result)

}
