% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\name{arln_process_dfs_to_row}
\alias{arln_process_dfs_to_row}
\title{Process a List of Data Frames and Pivot to Wide Format}
\usage{
arln_process_dfs_to_row(df_list)
}
\arguments{
\item{df_list}{A list of data frames. The function expects that the first
data frame in the list will serve as the base, and all subsequent data
frames will be processed and joined.}
}
\value{
A data frame in wide format, with columns based on the unique
values in the \code{"Keys"} column and corresponding values from the \code{"Values"}
column.
}
\description{
This function processes a list of data frames, applies the
\code{results_processing} function to each data frame (except the first one),
and then combines them into a single wide-format data frame. The function
performs the following operations:
\enumerate{
\item Applies the \code{results_processing} function to each data frame (excluding
the first one).
\item Combines the processed data frames into one using \code{bind_rows()}.
\item Pivots the resulting combined data frame to a wide format, with columns
derived from the \code{"Keys"} column and their corresponding values from the
\code{"Values"} column.
}
}
\examples{
df1 <- data.frame(Keys = c("gene1", "gene2"),
                  Data_A =c("a1", "a2"),
                  Values = c("bla_gene1", "bla_gene2"))
df2 <- data.frame(Keys = c("gene3", "gene4"),
                  Data_B =c("b1", "b2"),
                  Values = c("oxa_gene3", "oxa_gene4"))
df_list <- list(df1, df2)
result_df <- arln_process_dfs_to_row(df_list)
# result_df will be a wide-format data frame with columns
# for 'gene1', 'gene2', 'gene3', 'gene4'
# and the corresponding values from the 'Values' column

}
