% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{capitalize_column_names}
\alias{capitalize_column_names}
\title{Capitalize All Column Names}
\usage{
capitalize_column_names(df)
}
\arguments{
\item{df}{A data frame whose column names need to be capitalized.}
}
\value{
A data frame with column names where all letters have been
capitalized.
}
\description{
This function takes a data frame and capitalizes all the letters in the
column names.
It modifies the column names in-place and returns the updated data frame.
}
\examples{
df <- data.frame("name" = c("John", "Jane"), "age" = c(30, 25))
df_updated <- capitalize_column_names(df)
# df_updated will have column names "NAME" and "AGE"

}
