% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{clean_excel_column_names}
\alias{clean_excel_column_names}
\title{Clean Column Names by Replacing Newline and Carriage Return Characters}
\usage{
clean_excel_column_names(df, replace_with = "_")
}
\arguments{
\item{df}{A data frame whose column names will be cleaned.}

\item{replace_with}{A string to replace newline and carriage return
characters with (default is "_").}
}
\value{
A data frame with cleaned column names.
}
\description{
This function cleans the column names of a data frame by removing any newline
(\verb{\\n}) or carriage return (\verb{\\r}) characters.
The unwanted characters are replaced with a specified string (default is an
underscore \verb{_}).
}
\examples{
df <- data.frame("col1\r" = c(1, 2, 3), "col2\n" = c(4, 5, 6),
                 stringsAsFactors = FALSE)
df_cleaned <- clean_excel_column_names(df)
# The column names of df_cleaned will have "\r" and "\n" replaced by "_".

}
