% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\docType{data}
\name{COLUMN_NAMES}
\alias{COLUMN_NAMES}
\title{A list of column names used for data selection from Tenn ARLN pdfs.}
\format{
An object of class \code{character} of length 3.
}
\usage{
COLUMN_NAMES
}
\description{
A list of column names used for data selection from Tenn ARLN pdfs.
}
\examples{
COLUMN_NAMES <- c("Results",
                  "Reference Range",
                  "Performing Location")
}
\keyword{datasets}
