% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{convert_date_columns}
\alias{convert_date_columns}
\title{Convert Date Columns in a Data Frame}
\usage{
convert_date_columns(df, date_columns, date_format = "\%m-\%d-\%Y")
}
\arguments{
\item{df}{A data frame containing the columns to be converted.}

\item{date_columns}{A character vector specifying the names of the columns
to be converted to Date format.}

\item{date_format}{A string specifying the date format to be used for
conversion (default is "\%m-\%d-\%Y").}
}
\value{
A data frame with the specified columns converted to Date format.
}
\description{
This function converts specified columns in a data frame to Date format,
using a provided date format.
If the column exists in the data frame, it will be converted using the
as.Date` function.
If the column is not found, a warning message will be shown.
}
\examples{
df <- data.frame(name = c("Alice", "Bob"),
                 date_of_birth = c("01-15-1990", "12-22-1985"),
                 join_date = c("03-20-2015", "05-10-2017"),
                 stringsAsFactors = FALSE)
df_updated <- convert_date_columns(df, date_columns = c("date_of_birth",
                                   "join_date"))
# The 'date_of_birth' and 'join_date' columns will be converted to Date format.

}
