% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataframe.R
\name{create_cpo_seq}
\alias{create_cpo_seq}
\title{Process Files in the 'PdfCpoSeq' Directory}
\usage{
create_cpo_seq(PDF_CPO_SEQ_DIR)
}
\arguments{
\item{PDF_CPO_SEQ_DIR}{Directory for cpo seq pdf files.}
}
\value{
A data frame that combines the results of processing all files in the
'PdfCpoSeq' directory. If the directory doesn't exist or is empty, an error
will be raised.
}
\description{
This function checks if the 'PdfCpoSeq' directory exists inside the package's
'extdata' folder and if it contains any files. If files are found, it
processes them by extracting text from PDFs and then extracting specific
information from those texts using the \code{tenn_cpo_seq_extractor} function. The
results are then combined into a single data frame. If the directory does not
exist or is empty, an error is raised.
}
\details{
The function first checks if the 'PdfCpoSeq' directory exists within the
package's 'extdata' folder. If the directory is missing or contains no files,
it raises an error with an appropriate message. If files are present, it
processes each file using the \code{pdf_text} function from the \code{pdftools} package
to extract the text from the PDFs. Then, the \code{tenn_cpo_seq_extractor}
function is applied to the extracted text. Finally, the results are combined
using \code{bind_rows}.
}
