% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{create_event_dataframe_from_lines}
\alias{create_event_dataframe_from_lines}
\title{Create Event Dataframe from Lines}
\usage{
create_event_dataframe_from_lines(lines)
}
\arguments{
\item{lines}{A character vector where each element is a string containing a
key-value pair separated by a colon (\code{:}).}
}
\value{
A data frame with two columns:
\describe{
\item{Keys}{A character vector containing the keys extracted
from the input lines.}
\item{Values}{A character vector containing the values extracted
from the input lines.}
}
}
\description{
This function takes a vector of strings where each string is expected to
contain a key-value pair separated by a colon (\code{:}).
It processes each line, trims whitespace from the key and value, and
returns a data frame with two columns: "Keys" and "Values".
}
\examples{
# Example 1: Tenn ARLN pdf
# Extract the lines from the pdf
# Use system.file to get the path to the example document
  pdf_file_path <- system.file("extdata",
                               "pdf_tnarln_example_1.pdf",
                                package = "arlog")
  pdf <- pdftools::pdf_text(pdf_file_path)
# Convert the list of list of lists into a single list where each item is a
# line from the document.
  lines <- unlist(strsplit(pdf, "\n"))
  pdf_organism_culture <- search_for_culture_organism(lines,
                                                      CULTURE_ORGANISM_LIST)
  pdf_orgnsm_rslt <- create_event_dataframe_from_lines(pdf_organism_culture)
  print(pdf_orgnsm_rslt)
# Example 2:
lines <- c("Name: John Doe", "Age: 30", "Location: New York")
create_event_dataframe_from_lines(lines)

}
