% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataframe.R
\name{create_excel_cpo}
\alias{create_excel_cpo}
\title{Process Files in the 'ExcelCpo' Directory}
\usage{
create_excel_cpo(EXCEL_CPO_DIR)
}
\arguments{
\item{EXCEL_CPO_DIR}{Directory for excel cpo files.}
}
\value{
A data frame that combines the results of processing all files in the
'ExcelCpo' directory. If the directory doesn't exist or is empty, an error
will be raised.
}
\description{
This function checks if the 'ExcelCpo' directory exists within the package's
'extdata' folder and if it contains any files. If files are found, it processes
them using the \code{excel_cpo_extractor} function and combines the results into a
single data frame. If the directory does not exist or is empty, the function
raises an error.
}
\details{
The function first checks if the 'ExcelCpo' directory exists within the package's
'extdata' folder. If the directory is missing or contains no files, it raises an
error with an appropriate message. If files are present, it processes each file
using the \code{excel_cpo_extractor} function and combines the results into a single
data frame using \code{bind_rows}.
}
