% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\name{create_table_arln}
\alias{create_table_arln}
\title{Create a Unified Table with Standardized Columns}
\usage{
create_table_arln(rows)
}
\arguments{
\item{rows}{A list of data frames. Each data frame may have different
columns, and this function will standardize them by filling in missing
columns and combining them into one data frame.}
}
\value{
A data frame with standardized columns (based on the data frame with
the most columns) and trimmed character columns. Missing values are replaced
with \code{"Not in Report"}.
}
\description{
This function processes a list of data frames, ensuring they all have the
same columns by filling missing columns with a default value ("Not in
Report"). It then combines the data frames into one and trims any leading or
trailing white space from character columns. The resulting data frame is
returned.
}
\examples{
df1 <- data.frame(ColumnA = c(1, 2),
                  ColumnB = c(" A ", " B "),
                  stringsAsFactors = FALSE)
df2 <- data.frame(ColumnA = c(3, 4),
                  ColumnC = c(" C1 ", " C2 "),
                  stringsAsFactors = FALSE)
rows <- list(df1, df2)
final_df <- create_table_arln(rows)
# final_df will have columns "ColumnA", "ColumnB", and "ColumnC", with
# missing values filled
# and leading/trailing spaces trimmed from character columns.

}
