% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataframe.R
\name{create_tenn_arln}
\alias{create_tenn_arln}
\title{Create a Table from 'PdfArln' Directory Files}
\usage{
create_tenn_arln(PDF_ARLN_DIR)
}
\arguments{
\item{PDF_ARLN_DIR}{File directory for ARLN pdf files.}
}
\value{
A data frame containing the processed table generated from the files
in the 'PdfArln' directory. If the directory doesn't exist or is empty,
an error is raised.
}
\description{
This function processes PDF files located in the 'PdfArln' directory inside
the package's 'extdata' folder. It checks whether the directory exists and
contains any files. If files are found, they are processed and a table is
created from the data in those files. If the directory doesn't exist or is
empty, an error is raised.
}
\details{
The function checks if the 'PdfArln' directory exists within the package's
'extdata' folder. If the directory is missing or contains no files, it will
raise an error with an appropriate message. If files are present, the function
will read the files and process them using the \code{data_rows} and
\code{create_table_arln} functions to generate a final output table.
}
