% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataframe.R
\name{create_web_portal}
\alias{create_web_portal}
\title{Process Files in the 'ExcelArlnWebPortal' Directory}
\usage{
create_web_portal(EXCEL_ARLN_WEB_PORTAL_DIR)
}
\arguments{
\item{EXCEL_ARLN_WEB_PORTAL_DIR}{Directory for ARLN web portal excel files.}
}
\value{
A data frame that combines the results of processing all CSV files in
.  the 'ExcelArlnWebPortal' directory. If the directory doesn't exist or is
empty, an error will be raised.
}
\description{
This function checks if the 'ExcelArlnWebPortal' directory exists inside the
package's 'extdata' folder and if it contains any files. If files are found,
it processes them by reading CSV files and then combines the results into a
single data frame.
If the directory does not exist or is empty, an error is raised.
}
\details{
The function first checks if the 'ExcelArlnWebPortal' directory exists within
the package's 'extdata' folder. If the directory is missing or contains no
files, it raises an error with an appropriate message. If files are present,
it reads each CSV file using \code{read.csv} and combines the results using
\code{bind_rows} from the \code{dplyr} package. This allows the user to work with the
data in a combined data frame format.
}
