% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportDataframe.R
\name{create_word_alert}
\alias{create_word_alert}
\title{Process Files in the 'WordAlert' Directory}
\usage{
create_word_alert(WORD_ALERT_DIR)
}
\arguments{
\item{WORD_ALERT_DIR}{Directory for CRE word alerts.}
}
\value{
A data frame that combines the results of processing all files in the
'WordAlert' directory. If the directory doesn't exist or is empty, an error
will be raised.
}
\description{
This function checks for the presence of files in the 'WordAlert' directory,
located inside the package's 'extdata' folder. If files are found, it
processes each file using the \code{word_extraction} function and returns a
combined data frame with the results. If the directory doesn't exist or is
empty, the function raises
an error.
}
\details{
The function first checks if the 'WordAlert' directory exists within the
package's 'extdata' folder. If the directory is missing or contains no files,
it raises an error with an appropriate message. If files are present, it uses
the \code{word_extraction} function to process each file, and combines the results
into a single data frame using \code{bind_rows}.
}
