% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\docType{data}
\name{CULTURE_ORGANISM_LIST}
\alias{CULTURE_ORGANISM_LIST}
\title{A list of Results used for data selection from Tenn ARLN pdfs.}
\format{
An object of class \code{character} of length 2.
}
\usage{
CULTURE_ORGANISM_LIST
}
\description{
A list of Results used for data selection from Tenn ARLN pdfs.
}
\examples{
CULTURE_ORGANISM_LIST <- c("RESULT:Carbapenem Resistant Organism Culture",
                           "RESULT:Mycology Culture")
}
\keyword{datasets}
