% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{cut_off_time}
\alias{cut_off_time}
\title{Cut Off Time from DateTime Columns}
\usage{
cut_off_time(df, datetime_columns)
}
\arguments{
\item{df}{A data frame containing the datetime columns to be processed.}

\item{datetime_columns}{A character vector of column names in \code{df} that are datetime columns.
The function will truncate the time from these columns, keeping only the date part.}
}
\value{
A data frame with the specified datetime columns modified, where the time part is removed
and only the date is kept.
}
\description{
This function processes specified datetime columns in a data frame and truncates
the time portion, keeping only the date. It removes everything after the space
(which separates the date and time) for each specified column.
}
\examples{
# Sample data frame with datetime columns
df <- data.frame(
  ID = 1:3,
  created_at = c("2024-12-01 12:34:56", "2024-12-02 15:22:33", "2024-12-03 08:45:01"),
  updated_at = c("2024-12-01 12:34:56", "2024-12-02 15:22:33", "2024-12-03 08:45:01"),
  stringsAsFactors = FALSE
)

# Apply the cut_off_time function to truncate time part from 'created_at' and 'updated_at'
df_cleaned <- cut_off_time(df, c("created_at", "updated_at"))

# The resulting df_cleaned will have only the date part in 'created_at' and 'updated_at'

}
