% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\name{data_rows}
\alias{data_rows}
\title{Process a List of PDF Files and Extract Data}
\usage{
data_rows(file_list)
}
\arguments{
\item{file_list}{A character vector of file paths to the PDF files to be processed.}
}
\value{
A list of data frames, where each data frame corresponds to the data extracted
from a single PDF file and processed into rows.
}
\description{
This function processes a list of PDF files, extracting text from each PDF
using the \code{pdftools::pdf_text} function, then applies the \code{tenn_arln_pdf_extractor}
function to extract relevant data from the PDF content, and finally processes
the data frame using the \code{arln_process_dfs_to_row} function.
}
\examples{
example1 = system.file("extdata",
                       "pdf_tnarln_example_1.pdf",
                        package = "arlog")
example2 = system.file("extdata",
                       "pdf_tnarln_example_2.pdf",
                        package = "arlog")
file_list <- c(example1, example2)

data_rows(file_list)

}
