% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDocx.R
\name{docx_extract_table}
\alias{docx_extract_table}
\title{Extract the first table from a word document.}
\usage{
docx_extract_table(file_path)
}
\arguments{
\item{file_path}{The file path for a CRE Alert word document containing one table.}
}
\value{
The CRE Alert table from the word document.
}
\description{
Used in the package to extract word documents. Test using system.file.
}
\examples{
 # Use system.file to get the path to the example document
  word_file_path <- system.file("extdata", "word_example_document.docx", package = "arlog")

 # Read the document and extract the first table
  doc_table <- docx_extract_table(word_file_path)
  print(doc_table)

}
