% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDocx.R
\name{docx_rename_columns_with_first_row}
\alias{docx_rename_columns_with_first_row}
\title{Rename the columns of a table with the elements of the first row of the table.}
\usage{
docx_rename_columns_with_first_row(table)
}
\arguments{
\item{table}{The table extracted from a CRE alert by docx_extract_table.}
}
\value{
The table from the CRE alert with renamed columns.
}
\description{
Some docx files have a header row with two title column values irrelevant to the AR Log.
The first row contains the actual column names. This function takes the first data row, sets
it as column names and then removes this row of column names from the returned table.
}
\examples{
  word_file_path <- system.file("extdata", "word_example_document.docx", package = "arlog")
# Use docx_extract_table to read the document.
  doc_table <- docx_extract_table(word_file_path)

# Read the document and extract the first table
  renamed_table <- docx_rename_columns_with_first_row(doc_table)
  print(renamed_table)
}
