% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsXlsx.R
\name{excel_set_second_row_as_colnames}
\alias{excel_set_second_row_as_colnames}
\title{Rename the columns of a table with the elements of the first row of the table.}
\usage{
excel_set_second_row_as_colnames(data)
}
\arguments{
\item{data}{The table extracted from an excel file.}
}
\value{
The table from the excel file with the renamed columns.
}
\description{
Some excel files have a header row with two title column values irrelevant to the AR Log.
The first row contains the actual column names. This function takes the first data row,
sets it as column names and then removes this row of column names from the returned table.
}
\examples{
# Use system.file to to get the path to the example document
  excel_file_path <- system.file("extdata", "excel_example_document.xlsx", package = "arlog")
# Use openxlsx to read in an excel file
  exl <- openxlsx::read.xlsx(excel_file_path, colNames=TRUE)

# Use the function to return the correctly formatted table.
  excel_table <- excel_set_second_row_as_colnames(exl)
  print(excel_table)

}
