% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{export_data}
\alias{export_data}
\title{Export data to files.}
\usage{
export_data()
}
\value{
Returns nothing just writes files to a folder.
}
\description{
Export data to files.
}
\examples{
example text does a thing.. lalalala
}
