% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{extract_accession_info}
\alias{extract_accession_info}
\title{Extract Accession Information from Text}
\usage{
extract_accession_info(lines)
}
\arguments{
\item{lines}{A character vector (list of lines) containing the text from which
to extract the accession information.}
}
\value{
A data frame with two columns:
\itemize{
\item \code{Keys}: A character string representing the key "Accession #".
\item \code{Values}: A character string representing the accession number (value).
}
}
\description{
Some of the values in the columns cannot be extracted by using the key-value
matching form the previous function. Accession # is one of those values.
This function extracts the accession number information from a list of lines
of text.It looks for a line containing the phrase "Accession #" and splits
the line into a key-value pair, where the key is "Accession #" and the value
is the corresponding accession number.
}
\details{
The function searches through the input text and looks for a line containing
the phrase "Accession #". If the line is found, it splits the line at the
colon (\code{:}) to separate the key from the value. The function returns a data
frame with these key-value pairs. If no line containing "Accession #" is
found, the function throws an error.
}
\examples{
# Example 1: Tenn ARLN pdf
# Use system.file to get the path to the example document
  pdf_file_path <- system.file("extdata",
                               "pdf_tnarln_example_1.pdf",
                                package = "arlog")

# Extract the lines from the pdf
  pdf <- pdftools::pdf_text(pdf_file_path)
# Convert the list of list of lists into a single list where each item is a
# line from the document.
  lines <- unlist(strsplit(pdf, "\n"))
  result <- pdf_extract_lines_between(lines, "Accession", "Key")
  accession_num <- extract_accession_info(result)
  print(accession_num)

# Example 2: Extract accession information from a sample text
lines <- c("Header", "Accession #: ABC12345", "Footer")
result <- extract_accession_info(lines)
print(result)

# Example: Error case when "Accession #" is not found
lines <- c("Header", "Footer")
# This will throw an error:
# extract_accession_info(lines)

}
