% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{extract_key_value}
\alias{extract_key_value}
\title{Extract Key-Value Pair from String}
\usage{
extract_key_value(cell)
}
\arguments{
\item{cell}{A string representing a key-value pair in the format
"key: value".
If the string is empty or contains "\if{html}{\out{<NA>}}", the function
returns \code{NULL}.}
}
\value{
A list with two elements:
\item{key}{The part before the colon (":") as the key.}
\item{value}{The part after the colon as the value. If no value is
found, \code{NA} is returned.}
If the input is invalid (i.e., \code{NA} or "\if{html}{\out{<NA>}}"), the function
returns \code{NULL}.
}
\description{
Function to extract keys and values from a colon-separated string. This
function is intended to be used by other functions and is not called
directly on dataframes. It splits a string at the first occurrence of
\code{": "} and returns a list containing the key and value. If the input
string is not in the expected format or is missing, it returns \code{NULL} or
a list with a key and \code{NA} for the value.
}
\examples{
extract_key_value("name: John Doe")
# Returns: list(key = "name", value = "John Doe")

extract_key_value("<NA>")
# Returns: NULL

extract_key_value("age: ")
# Returns: list(key = "age", value = NA)

}
