% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\name{fill_missing_columns}
\alias{fill_missing_columns}
\title{Fill Missing Columns with Default Values}
\usage{
fill_missing_columns(df, all_columns)
}
\arguments{
\item{df}{A data frame that may have a subset of the columns defined in
\code{all_columns}.}

\item{all_columns}{A character vector of column names that should be present
in the output data frame.}
}
\value{
A data frame that contains all columns specified in \code{all_columns}.
Missing columns are added and filled with the value \code{"Not in Report"}.
Existing columns from the input data frame are retained.
}
\description{
This function ensures that a data frame contains all specified columns. If
any columns are missing, they are added to the data frame and filled with the
default value \code{"Not in Report"}. Existing columns from the input data frame
are retained. This function is useful for standardizing data frames to a
common set of columns.
}
\examples{
# Example data frames (df1 and df2) to demonstrate the function
df1 <- data.frame(Keys = c("gene1", "gene2"),
                  Data_A =c("a1", "a2"),
                  Values = c("bla_gene1", "bla_gene2"))
df2 <- data.frame(Keys = c("gene3", "gene4"),
                  Data_B =c("b1", "b2"),
                  Values = c("oxa_gene3", "oxa_gene4"))

# Combine the data frames into a list
df_list <- list(df1, df2)

# Process the data frames and convert to wide format
result_df <- arln_process_dfs_to_row(df_list)

# View the resulting wide-format data frame
print(result_df)
# The result will be a wide-format data frame with columns:
# 'gene1', 'gene2', 'gene3', 'gene4' and their corresponding values.

}
