% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{pdf_extract_lines_between}
\alias{pdf_extract_lines_between}
\title{Extract lines between two words from a list of lines.}
\usage{
pdf_extract_lines_between(lines, start_string, end_string)
}
\arguments{
\item{lines}{The lines extracted from reading the pdf or other document.}

\item{start_string}{The word that starts the line selection, inclusive.}

\item{end_string}{The word that ends the line selection, not inclusive.}
}
\value{
The lines between start and stop words, including start line and
exlcuding the stop line.
}
\description{
In the TN Department of Health PDFs, exporting the data requires reading the
pdf and extracting different segments at a time. This function gives a
selection of lines from a list of lines based on a start and stop word.
}
\examples{
# Use system.file to get the path to the example document
  pdf_file_path <- system.file("extdata",
                               "pdf_tnarln_example_1.pdf",
                                package = "arlog")

# Extract the lines from the pdf
  pdf <- pdftools::pdf_text(pdf_file_path)
# Convert the list of list of lists into a single list where each item is a
# line from the document.
  lines <- unlist(strsplit(pdf, "\n"))
  result <- pdf_extract_lines_between(lines, "Accession", "Key")
  print(result)
}
