% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{pdf_split_lines_into_columns}
\alias{pdf_split_lines_into_columns}
\title{Split lines into columns when spacing between words is larger than 9}
\usage{
pdf_split_lines_into_columns(lines)
}
\arguments{
\item{lines}{The lines extracted from a Tenn ARLN pdf between the words
"Accession" and "Key".}
}
\value{
A data frame with three columns, Column1, Column2, Column3.
}
\description{
The Tennessee Dept of Health PDFs feature a Preliminary report section. The
section is divided visually into three columns. This function takes the lines
extracted from the document and breaks the lines into three columns based on
the distance between the words, when there are more than 9 spaces between
words, that marks a column break.
}
\examples{
# Use system.file to get the path to the example document
  pdf_file_path <- system.file("extdata",
                               "pdf_tnarln_example_1.pdf",
                                package = "arlog")

# Extract the lines from the pdf
  pdf <- pdftools::pdf_text(pdf_file_path)
# Convert the list of list of lists into a single list where each item is a
# line from the document.
  lines <- unlist(strsplit(pdf, "\n"))
  result <- pdf_extract_lines_between(lines, "Accession", "Key")
# Convert the selected lines into a dataframe of three columns.
  result_df <- pdf_split_lines_into_columns(result)
  print(result_df)
}
