% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\name{prefix_column_single_by_test}
\alias{prefix_column_single_by_test}
\title{Add Prefix to Specific Columns Based on Name}
\usage{
prefix_column_single_by_test(data)
}
\arguments{
\item{data}{A data frame that may contain the columns \code{"beta-lactamase"} or
\code{"OXA"}. The function will add the respective prefix to the values in these
columns if they are present.}
}
\value{
The modified data frame with prefixed values in the
\code{"beta-lactamase"} or \code{"OXA"} columns.
}
\description{
This function checks if certain columns (\code{"beta-lactamase"} or \code{"OXA"})
exist in a given data frame. If the \code{"beta-lactamase"} column is present,
the function adds the prefix \code{"bla_"} to each value in that column. If the
\code{"OXA"} column exists instead, it adds the prefix \code{"oxa_"}. The modified
data frame is then returned.
}
\examples{
df <- data.frame(`beta-lactamase` = c("gene1", "gene2"),
                  OXA = c("geneA", "geneB"),
                  stringsAsFactors = FALSE)
result_df <- prefix_column_single_by_test(df)
# result_df will have "bla_gene1", "bla_gene2" in the 'beta-lactamase'
# column and "oxa_geneA", "oxa_geneB" in the 'OXA' column

}
