% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\docType{data}
\name{PREFIX_LIST}
\alias{PREFIX_LIST}
\title{A list of prefixes to associate data with the respective test}
\format{
An object of class \code{character} of length 2.
}
\usage{
PREFIX_LIST
}
\description{
A list of prefixes to associate data with the respective test
}
\examples{
PREFIX_LIST <- c("oxa","bla")
}
\keyword{datasets}
