% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCsv.R
\name{process_excel_cpo}
\alias{process_excel_cpo}
\title{Process Excel CPO Data}
\usage{
process_excel_cpo(cpo_df, CSV_OUTPUT_DIR)
}
\arguments{
\item{cpo_df}{A DataFrame containing the raw CPO data from an Excel sheet.}

\item{CSV_OUTPUT_DIR}{Directory to place the created csv file.}
}
\value{
NULL This function does not return any value. It saves the processed
DataFrame to a CSV file.
}
\description{
This function processes raw CPO data from an Excel file by
cleaning the column names, converting date columns to the correct format, and
saving the processed data to a CSV file in a specified directory.
}
\details{
The function performs the following operations on the input DataFrame:
\itemize{
\item Cleans Excel column names (removes carriage returns, newlines, and
converts characters).
\item Capitalizes all column names.
\item Replaces "DOB" with "DATE_OF_BIRTH" and capitalizes.
\item Replaces double underscores in column names with single underscores.
\item Replaces spaces in column names with underscores.
\item Converts date columns to the correct date format (MM/DD/YYYY).
}

It then saves the processed data to a CSV file in a directory specified by
the global variable \code{CSV_OUTPUT_DIR} with the filename \code{"excel_cpo.csv"}.
}
