% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCsv.R
\name{process_excel_sentinel}
\alias{process_excel_sentinel}
\title{Process Excel Sentinel Data}
\usage{
process_excel_sentinel(df, CSV_OUTPUT_DIR)
}
\arguments{
\item{df}{A DataFrame containing the raw Sentinel data from an Excel sheet.}

\item{CSV_OUTPUT_DIR}{Directory to place the created csv file.}
}
\value{
NULL This function does not return any value. It saves the processed
DataFrame to a CSV file.
}
\description{
This function processes raw Sentinel data from an Excel file,
specifically converting certain date columns to the correct date format and
saving the processed data to a CSV file in a specified directory.
}
\details{
The function performs the following operations on the input
DataFrame:
\itemize{
\item Replaces "DOB" with "DATE_OF_BIRTH" and ensures correct date format.
\item Converts the columns \code{DATE_OF_BIRTH}, \code{CULTUREDT}, and \code{COLLECTIONDT}
to the correct date format (MM/DD/YYYY).
}

It then saves the processed data to a CSV file in a directory specified by
the global variable \code{CSV_OUTPUT_DIR} with the filename \code{"excel_sent.csv"}.
}
