% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCsv.R
\name{process_pdf_cpo_seq}
\alias{process_pdf_cpo_seq}
\title{Process PDF CPO Sequence Data}
\usage{
process_pdf_cpo_seq(df, CSV_OUTPUT_DIR)
}
\arguments{
\item{df}{A DataFrame containing the raw CPO sequence data extracted from a
PDF.}

\item{CSV_OUTPUT_DIR}{Directory to place the created csv file.}
}
\value{
NULL This function does not return any value. It saves the processed
DataFrame to a CSV file.
}
\description{
This function processes raw PDF CPO sequence data, specifically
converting the \code{COLLECTION_DATE} column to the correct date format
(MM/DD/YYYY) and saving the processed data to a CSV file in a specified
directory.
}
\details{
The function performs the following operations on the input
DataFrame:
\itemize{
\item Capitalizes the column names to ensure uniformity.
\item Converts the \code{COLLECTION_DATE} column to the correct date format
(MM/DD/YYYY).
}

It then saves the processed data to a CSV file in a directory specified by
the global variable \code{CsvOutputDir} with the filename \code{"pdf_cpo.csv"}.
}
