% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCsv.R
\name{process_tenn_arln}
\alias{process_tenn_arln}
\title{Process the 'tenn_arln' Data Frame and Save as CSV}
\usage{
process_tenn_arln(tenn_arln_df, CSV_OUTPUT_DIR)
}
\arguments{
\item{tenn_arln_df}{A \code{tenn_arln} data frame to be processed}

\item{CSV_OUTPUT_DIR}{Directory to place the created csv file.}
}
\value{
NULL
This function does not return anything. It saves the processed data as a CSV file
in the \code{CSV_OUTPUT_DIR} directory.
}
\description{
This function processes the \code{tenn_arln_df} data frame by applying a series of
transformations:
\enumerate{
\item Replaces spaces in column names with underscores.
\item Replaces any \verb{#} characters in column names with the string "Num".
\item Capitalizes all column names.
}

After performing these transformations, the processed data is saved as a CSV
file in the specified output directory.
}
\details{
The function expects the \code{CSV_OUTPUT_DIR} variable to specify the directory
where the resulting CSV file will be saved. The output file will be named
\code{tenn_arln.csv}.The function does not return any output to the user, but it
will write the processed data frame to the specified location.
}
