% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportCsv.R
\name{process_word_alert}
\alias{process_word_alert}
\title{Process the 'word_alert' Data Frame and Save as CSV}
\usage{
process_word_alert(word_alert_df, CSV_OUTPUT_DIR)
}
\arguments{
\item{word_alert_df}{A data frame to be processed, containing the
\code{word_alert} data.}

\item{CSV_OUTPUT_DIR}{Directory to place the created csv file.}
}
\value{
NULL
This function does not return anything. It saves the processed data as a
CSV file in the \code{Processed} directory.
}
\description{
This function processes the \code{word_alert_df} data frame by applying a series of
transformations:
\enumerate{
\item Replaces spaces in column names with underscores.
\item Capitalizes all column names.
\item Replaces "DOB" with "DATE_OF_BIRTH" in column names.
\item Replaces hyphens (more than 2 consecutive hyphens) in character columns
with \code{NA}.
\item Converts date columns ("DATE_OF_BIRTH" and "DATE_OF_COLLECTION") from
character to Date type using the specified date format (\verb{\%m-\%d-\%Y}).
}

After performing these transformations, the processed data is saved as a CSV
file in the specified output directory.
}
\details{
The function expects the \code{CSV_OUTPUT_DIR} variable to specify the directory
where the resulting CSV file will be saved. The output file will be named
\code{word_alert.csv}. The function does not return any output to the user,
but it will write the processed data frame to the specified location.
}
