% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{rename_dataframe_columns}
\alias{rename_dataframe_columns}
\title{Rename Data Frame Columns}
\usage{
rename_dataframe_columns(df, column_names)
}
\arguments{
\item{df}{A data frame whose columns will be renamed.}

\item{column_names}{A character vector of new column names. The length of
this vector must match the number of columns in the \code{df}.}
}
\value{
The data frame with updated column names.
}
\description{
This function renames the columns of a data frame by setting the column
names to the values provided in the \code{column_names} vector. It checks if the
length of \code{column_names} matches the number of columns in the data frame.
If the lengths do not match, an error is raised.
}
\examples{
df <- data.frame(A = 1:3, B = 4:6)
new_column_names <- c("Column1", "Column2")
df_renamed <- rename_dataframe_columns(df, new_column_names)
print(df_renamed)

}
