% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{replace_dob_with_date_of_birth}
\alias{replace_dob_with_date_of_birth}
\title{Replace "DOB" with "DATE_OF_BIRTH" in Column Names}
\usage{
replace_dob_with_date_of_birth(df)
}
\arguments{
\item{df}{A data frame whose column names will be modified.}
}
\value{
A data frame with the column names modified such that any "DOB"
entries are replaced with "DATE_OF_BIRTH",
and all letters are capitalized.
}
\description{
This function replaces the string "DOB" with "DATE_OF_BIRTH" in the column
names of the input data frame.
It also capitalizes all column names to ensure consistency.
}
\examples{
df <- data.frame("DOB" = c("1990-01-01", "1985-05-20"),
                 "name" = c("John", "Jane"))
df_updated <- replace_dob_with_date_of_birth(df)
# df_updated will have column names "DATE_OF_BIRTH" and "NAME"

}
