% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{replace_hash_with_num}
\alias{replace_hash_with_num}
\title{Replace Hashes with 'Num' in Column Names}
\usage{
replace_hash_with_num(df)
}
\arguments{
\item{df}{A data frame whose column names need to be modified.}
}
\value{
A data frame with column names where all occurrences of \verb{#} have been
replaced by "Num".
}
\description{
This function takes a data frame and replaces all occurrences of the hash
symbol (\verb{#}) in the column names
with the string "Num". It modifies the data frame in-place and returns the
updated data frame.
}
\examples{
df <- data.frame("#ID" = 1:3, "Age#" = c(25, 30, 35))
df_updated <- replace_hash_with_num(df)
# df_updated will have columns "NumID" and "AgeNum"

}
