% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{replace_hyphens_with_na}
\alias{replace_hyphens_with_na}
\title{Replace Hyphens with NA in Data Frame}
\usage{
replace_hyphens_with_na(df)
}
\arguments{
\item{df}{A data frame where hyphens in character columns will be replaced
by \code{NA}.}
}
\value{
A data frame with the specified hyphen strings replaced by \code{NA}.
}
\description{
This function searches for cells in the data frame that contain a string of
hyphens ("-") of length greater than 2
and replaces those occurrences with \code{NA}. The function applies this
replacement to all columns in the data frame
where the columns contain character data.
}
\examples{
df <- data.frame(col1 = c("abc", "---", "xyz"), col2 = c("123", "--", "---"))
df_updated <- replace_hyphens_with_na(df)
# df_updated will have `NA` in place of hyphen strings like "---"

}
