% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsDataProcessing.R
\name{replace_spaces_with_underscores}
\alias{replace_spaces_with_underscores}
\title{Replace Spaces with Underscores in Column Names}
\usage{
replace_spaces_with_underscores(df)
}
\arguments{
\item{df}{A data frame whose column names need to be modified.}
}
\value{
A data frame with column names where spaces are replaced by
underscores.
}
\description{
This function takes a data frame and replaces all spaces in the column names
with underscores. It modifies the data frame in-place and returns the updated
data frame.
}
\examples{
df <- data.frame("Column One" = 1:3, "Column Two" = 4:6)
df_updated <- replace_spaces_with_underscores(df)
# df_updated will have columns "Column_One" and "Column_Two"

}
