% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\name{results_processing}
\alias{results_processing}
\title{Process Results for Join Operation}
\usage{
results_processing(df)
}
\arguments{
\item{df}{A data frame to be processed. The function expects the input data
frame to have columns that may require prefixing and renaming.}
}
\value{
A data frame that has been processed with the appropriate prefix
added to columns and the first and third columns renamed to \code{"Keys"} and
\code{"Values"}, respectively.
}
\description{
This function processes the input data frame by applying two operations:
\enumerate{
\item It adds a prefix to specific columns (using the
\code{prefix_column_single_by_test} function).
\item It selects and renames the first and third columns (using the
\code{rename_columns_for_join} function) to prepare them for a join operation.
The function returns the modified data frame with the appropriate
transformations applied.
}
}
\examples{
df <- data.frame(`beta-lactamase` = c("IMP", "KPC", "OXA-48","VIM","NDM"),
                  Results = c("not det",
                              "not det",
                              "not det",
                              "not det",
                              "det"),
                  `Reference Range` = c("not det",
                                        "not det",
                                        "not det",
                                        "not det",
                                        "not det" ),
                  Location = c("NYC", "NYC", "NYC", "NYC", "NYC"),
                  stringsAsFactors = FALSE
                  )
processed_df <- results_processing(df)
# processed_df will have the `"beta-lactamase"` or `"OXA"` columns prefixed,
# and the first and third columns renamed to 'Keys' and 'Values'.

}
