% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\name{search_for_culture_organism}
\alias{search_for_culture_organism}
\title{Search for Culture Organism in Lines}
\usage{
search_for_culture_organism(lines, CULTURE_ORGANISM_LIST)
}
\arguments{
\item{lines}{A character vector containing lines of text to search through.}

\item{CULTURE_ORGANISM_LIST}{A character vector containing the names of
organisms to search for within the lines.}
}
\value{
A list of lines from the input \code{lines} that contain any of the
organisms in \code{CULTURE_ORGANISM_LIST}.
}
\description{
This function searches through a list of lines (e.g., text lines) and returns
the lines that contain any of the organisms specified in the provided list
\code{CULTURE_ORGANISM_LIST}. It performs a case-insensitive search for the
organisms within each line.
}
\examples{
# Example 1: Tenn ARLN pdf
# Use system.file to get the path to the example document
  pdf_file_path <- system.file("extdata",
                               "pdf_tnarln_example_1.pdf",
                                package = "arlog")
# Extract the lines from the pdf
  pdf <- pdftools::pdf_text(pdf_file_path)
# Convert the list of list of lists into a single list where each item is a
# line from the document.
  lines <- unlist(strsplit(pdf, "\n"))
  pdf_organism_culture <- search_for_culture_organism(lines,
                                                      CULTURE_ORGANISM_LIST)
  print(pdf_organism_culture)
# Example 2:
lines <- c("The culture of E. coli was successful.",
           "Yeast culture did not grow.",
           "Salmonella growth observed.")
CULTURE_ORGANISM_LIST <- c("E. coli", "Yeast", "Salmonella")
search_for_culture_organism(lines, CULTURE_ORGANISM_LIST)

}
