% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData.R
\name{tenn_arln_pdf_extractor}
\alias{tenn_arln_pdf_extractor}
\title{Convert ARLN pdf into three tables of data}
\usage{
tenn_arln_pdf_extractor(pdf)
}
\arguments{
\item{pdf}{The lines read in from a pdf file.}
}
\value{
A list of tables.
}
\description{
The function converts one Tennessee ARLN pdf into three tables. One
containing the patient/culture information from the first page, and
additional tables for each of the reported results on the last page
of the report.
}
\examples{
# Read in pdf and extract lines
pdf_file_path <- system.file("extdata",
                             "pdf_tnarln_example_1.pdf",
                             package = "arlog")
pdf <- pdftools::pdf_text(pdf_file_path)
results <- tenn_arln_pdf_extractor(pdf)
print(results[[1]])
print(results[[2]])
}
