% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCpoSeqReport.R
\name{tenn_cpo_seq_extractor}
\alias{tenn_cpo_seq_extractor}
\title{Extracts a dataframe from CPO Sequencing Report}
\usage{
tenn_cpo_seq_extractor(pdf)
}
\arguments{
\item{pdf}{pdf lines read in from pdftools.}
}
\value{
Dataframe with data extracted from CPO pdf.
}
\description{
Extracts the tables found on the second page of the CPO sequencing
report pdf.
}
\examples{
# Read in pdf and extract lines
pdf_file_path <- system.file("extdata",
                             "pdf_cposeq_example.pdf",
                             package = "arlog")
pdf <- pdftools::pdf_text(pdf_file_path)
df <- tenn_cpo_seq_extractor(pdf)
print(df)
}
