% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTennPdfData2.R
\docType{data}
\name{TEST_LIST}
\alias{TEST_LIST}
\title{A list of test types that appear in ARLN pdfs}
\format{
An object of class \code{character} of length 2.
}
\usage{
TEST_LIST
}
\description{
A list of test types that appear in ARLN pdfs
}
\examples{
TEST_LIST <- c("OXA","beta-lactamase")
}
\keyword{datasets}
