% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToolsPdfExtraction.R
\docType{data}
\name{VALUES_TO_KEEP}
\alias{VALUES_TO_KEEP}
\title{A list of columns to keep from the first page of the Tenn ARLN pdfs.}
\format{
An object of class \code{character} of length 6.
}
\usage{
VALUES_TO_KEEP
}
\description{
A list of columns to keep from the first page of the Tenn ARLN pdfs.
}
\examples{
VALUES_TO_KEEP <- c("Patient Name",
                    "Date of Birth",
                    "Specimen Type",
                    "Date Collected",
                    "Date Received",
                    "Date Reported")
}
\keyword{datasets}
