% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDocx.R
\name{word_extraction}
\alias{word_extraction}
\title{Full process for extracting a table correctly from a CRE alert.}
\usage{
word_extraction(word_file_path)
}
\arguments{
\item{word_file_path}{The file path for a CRE Alert word document containing one table.}
}
\value{
The table from a CRE alert with columns headers extracted from the first data row.
}
\description{
This function is called through the main function in the program. It
is the full extraction process for a CRE Alert that comes in as a word
document, in the WordAlert folder. The function calls its functions from
ToolsDocx.R.
}
\examples{
# Use system.file to to get the path to the example document
  word_file_path <- system.file("extdata", "word_example_document.docx", package = "arlog")
  word_table <- word_extraction(word_file_path)
  print(word_table)

}
